Add failing test case for having a dependency in multiple scopes.
In our case, we use logback-common in the runtime scope (production code is compiled against
com.typesafe/scalalogging-slf4j_2.10 scalalogging-slf4j_2.10-1.0.1.jar). We also use
logback-common in the test scope. We depend on the implementation for testing logging.

mpeltonen/sbt-idea#233 changed the ordering so that the jar is only in the Runtime scope in IntelliJ.
This prevents IntelliJ from compiling the tests.

I hacked around with the modules manually, and having the jar listed twice in IntelliJ module dependencies
  in multiple scopes makes it work for the manual case I tried. Specifically, this is the main.iml change.

I believe the checked in test will make my scenario work. The only difference is I manually set the
  "SBT: ch.qos.logback:logback-core:1.0.13:runtime" to be in the Test scope in IntelliJ. I'm guessing having
  two different references to the jar will work too.